+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Changelog
 ### 1.3.0
 - Started keeping a log of all changes between releases
 - Added support for quickly exposing functions as cli clients with `hug.cli` decorator
-- Added support for quickly serving up development APIs from withing the module using: `if __name__ == '__main__': __hug__.serve()
+- Added support for quickly serving up development APIs from withing the module using: `if __name__ == '__main__': __hug__.serve()`
 - Added support for documentation only type annotations, simply by passing just a string in as the type annotation
 - Added support for `requires` argument to limit execution of functions based on a given criteria
 - Added automatic documentation of output type transformations
@@ -29,4 +29,5 @@ Changelog
 ### 1.5.0
 - Added built-in support for outputting svg images
 - Added support for rendering images from pygal graphs, or other image framworks that support `render`, automatically
-- Added support for marshmellow powered output transformations
+- Added support for marshmallow powered output transformations
+- Added support for marshmallow powered input types